+++ jacsstorage-core/src/main/java/org/janelia/jacsstorage/io/TarArchiveBundleWriter.java
@@ -21,6 +21,9 @@ public class TarArchiveBundleWriter extends AbstractBundleWriter {
     @Override
     protected long writeBundleBytes(InputStream stream, String target) throws Exception {
         Path targetPath = Paths.get(target);
+        if (Files.exists(targetPath)) {
+            throw new IllegalArgumentException("Target path " + target + " already exists");
+        }
         Files.createDirectories(targetPath.getParent());
         return Files.copy(stream, Paths.get(target));
     }
